# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PRM Marketplace API",
    "description" : "Backend API for the student/resident marketplace",
    "contact" : {
      "name" : "PRM Marketplace Team"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/addresses/{id}" : {
      "get" : {
        "tags" : [ "address-controller" ],
        "operationId" : "read_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Address"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "address-controller" ],
        "operationId" : "update_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Address"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Address"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "address-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/auth/verify" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "verify",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerifyCodeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/reset-password" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "resetPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/resend-code" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "resendCode",
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/register" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/forgot-password" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "forgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/change-password" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "changePassword",
        "parameters" : [ {
          "name" : "currentPassword",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "newPassword",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/addresses" : {
      "get" : {
        "tags" : [ "address-controller" ],
        "operationId" : "getAll_9",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Address"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "address-controller" ],
        "operationId" : "create_10",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Address"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Address"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/addresses/{id}/default" : {
      "patch" : {
        "tags" : [ "address-controller" ],
        "operationId" : "setDefault",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Address"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "STUDENT", "FACULTY", "VENDOR", "RESIDENT" ]
          },
          "phone" : {
            "type" : "string"
          },
          "avatarUrl" : {
            "type" : "string"
          },
          "verified" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Address" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "line1" : {
            "type" : "string"
          },
          "line2" : {
            "type" : "string"
          },
          "suburb" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "postalCode" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "defaultAddress" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "singleLine" : {
            "type" : "string"
          }
        }
      },
...
      "VerifyCodeRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "code" : {
            "type" : "string",
            "maxLength" : 6,
            "minLength" : 6,
            "pattern" : "^[0-9]{6}$"
          }
        },
        "required" : [ "code", "email" ]
      },
      "ResetPasswordRequest" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string",
            "minLength" : 1
          },
          "newPassword" : {
            "type" : "string",
            "maxLength" : 2147483647,
            "minLength" : 8
          }
        },
        "required" : [ "newPassword", "token" ]
      },
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "maxLength" : 2147483647,
            "minLength" : 8
          },
          "role" : {
            "type" : "string",
            "enum" : [ "STUDENT", "FACULTY", "VENDOR", "RESIDENT" ]
          },
          "phone" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "name", "password" ]
      },
      "AuthResponse" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "email", "password" ]
      },
      "ForgotPasswordRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          }
        },
        "required" : [ "email" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}